Move get_mqtt_data to util.py
